Move Jormungandr API types into separate module.
