+++ lib/jormungandr/cardano-wallet-jormungandr.cabal
@@ -75,6 +75,7 @@ library
       Cardano.Wallet.Jormungandr
       Cardano.Wallet.Jormungandr.Api
       Cardano.Wallet.Jormungandr.Api.Client
+      Cardano.Wallet.Jormungandr.Api.Types
       Cardano.Wallet.Jormungandr.Binary
       Cardano.Wallet.Jormungandr.Compatibility
       Cardano.Wallet.Jormungandr.Network
